The first set of optimizations of Ampere, which changed the throughput of the FP64 significantly.


- [x] Remove CUB download as CUDA 11 now provides for it.
- [x] Use single precision cutoff when doing distance calculations.
- [x] Use a faster implementation of static_cast<long long>(float) for nonbonded kernels to improve performance 
- [x] Fixes a __syncwarp() issue in the neighborlist, oddly enough this only manifested in -g and -G builds
- [x] Clean up how we do the exclusions using the new fixed point conversion script

There is still roughly an 11% FP64 utilization coming from loading the double precision coordinates. We can address this either by buffering the nonbonded kernel, or just building a 32-bit context. But this will be done in a separate PR. Note that the -rbfe test have improved by ~60% 

```
Before
dhfr-apo: N=23558 speed: 217.05ns/day dt: 1.5fs (ran 100000 steps in 59.71s)
hif2a-apo: N=8805 speed: 480.90ns/day dt: 1.5fs (ran 100000 steps in 26.95s)
hif2a-rbfe: N=8840 speed: 243.46ns/day dt: 1.5fs (ran 100000 steps in 53.24s)
solvent-apo: N=6282 speed: 623.06ns/day dt: 1.5fs (ran 100000 steps in 20.80s)
solvent-rbfe: N=6317 speed: 317.68ns/day dt: 1.5fs (ran 100000 steps in 40.80s)

After
dhfr-apo: N=23558 speed: 244.34ns/day dt: 1.5fs (ran 100000 steps in 53.04s)
hif2a-apo: N=8805 speed: 486.00ns/day dt: 1.5fs (ran 100000 steps in 26.67s)
hif2a-rbfe: N=8840 speed: 389.50ns/day dt: 1.5fs (ran 100000 steps in 33.28s)
solvent-apo: N=6282 speed: 608.51ns/day dt: 1.5fs (ran 100000 steps in 21.30s)
solvent-rbfe: N=6317 speed: 494.53ns/day dt: 1.5fs (ran 100000 steps in 26.21s)
```